TwinMgr stats added and housekeeping for cache

+++ src/mqttRouterState.py
@@ -8,17 +8,22 @@ import paho.mqtt.client as mqtt
 from mqttRouter import MQTTRouter 
 import mqttTopicHelper as topicHelper
 from twinState import TwinState
+from twinObserver import TwinObserver
 import twinProtocol
+from mqttStateObserver import MQTTStateObserver
 import json
 
-class MQTTRouterState(MQTTRouter):
-    def __init__(self, client_id: str, state: TwinState):
+class MQTTRouterState(MQTTRouter, TwinObserver):
+    def __init__(self, client_id: str, state: TwinState, interface: mqtt):
         super().__init__(client_id)
         self.xLogging = logging.getLogger(__name__)
         self.xSet = topicHelper.getThingSet(self.getClientId())
         self.xGet = topicHelper.getThingGet(self.getClientId())
         self.xUpdate = topicHelper.getThingUpdate(self.getClientId())
         self.xState = state
+        self.xState.attachObserver(self)
+        self.interface = interface
+        self.xObservers  = []
        
     def subscribe(self, interface: mqtt):
         interface.subscribe(self.xGet, qos=1)
@@ -36,11 +41,15 @@ class MQTTRouterState(MQTTRouter):
         if (topic == self.xSet):
             j = json.loads(payload)
             if twinProtocol.TWINDELTA in j:
-                self.xState.updateState(j[twinProtocol.TWINDELTA])
+                d = self.setFilter(j[twinProtocol.TWINDELTA])
+                self.xState.updateState(d)
+                self.notifyObservers()
             if twinProtocol.TWINSTATE in j:
-                self.xState.deleteState()
-                self.xState.updateState(j[twinProtocol.TWINSTATE])
-                
+                if (self.deleteAllowed(topic, payload)):
+                    self.xState.deleteState()
+                d = self.setFilter(j[twinProtocol.TWINSTATE])
+                self.xState.updateState(d)
+                self.notifyObservers()
             self.sendUpdate(interface)
         
             self.xLogging.debug("Sucessfully Routed %s"%topic)
@@ -54,3 +63,34 @@ class MQTTRouterState(MQTTRouter):
         res[twinProtocol.TWINSTATE] = self.xState.getState()
         s = json.dumps(res)
         interface.publish(self.xUpdate, s, retain=False, qos=1)
+        
+        
+    def setFilter(self, delta: dict):
+        return delta
+    
+    def deleteAllowed(self, topic: str, payload: str):
+        return True
+    
+    def attachObserver(self, obs: MQTTStateObserver):
+        self.xObservers.append(obs)
+        
+    def dettachObserver(self, obs: MQTTStateObserver):
+        self.xObservers.remove(obs)    
+        
+    def notifyObservers(self):
+        for obs in self.xObservers:
+            obs.stateNotify(self.xState)  
+            
+            
+    def updateState(self, delta: dict, interface: mqtt = None):
+        self.xState.updateState(delta)
+        self.notifyObservers()
+        if (interface):
+            self.sendUpdate(interface)
+        
+    def getState(self):
+        return self.xState.getState()
+    
+    def notify(self):
+        self.sendUpdate(self.interface)
+
